# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebBanSachOnline API",
    "version": "1.0.0",
    "description": "Tài liệu API cho dự án WebBanSachOnline"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/auth/check-email": {
      "post": {
        "summary": "Kiểm tra email có tồn tại trong hệ thống không",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trả về { exists: true/false }",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": { "description": "Thiếu email" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "summary": "Gửi OTP đặt lại mật khẩu về email",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP đã được gửi (hoặc email không tồn tại — không tiết lộ)" },
          "429": { "description": "Gửi quá nhanh, cần chờ cooldown 60 giây" }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "summary": "Xác minh OTP 6 số, nhận reset token",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "otp"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "otp":   { "type": "string", "minLength": 6, "maxLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP đúng, trả về { ok, resetToken }" },
          "400": { "description": "OTP sai hoặc hết hạn" },
          "429": { "description": "Quá nhiều lần thử (tối đa 5 lần)" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "summary": "Đặt lại mật khẩu mới sau khi xác minh OTP",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "resetToken", "newPassword"],
                "properties": {
                  "email":       { "type": "string", "format": "email" },
                  "resetToken":  { "type": "string", "description": "Token nhận được từ /auth/verify-otp" },
                  "newPassword": { "type": "string", "minLength": 8, "description": "Tối thiểu 8 ký tự, gồm chữ hoa, chữ thường, số và ký tự đặc biệt" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Mật khẩu đã được cập nhật thành công" },
          "400": { "description": "Token không hợp lệ, hết hạn, hoặc mật khẩu yếu" },
          "404": { "description": "Tài khoản không tồn tại" }
        }
      }
    },
...
  }
}